Starting new plugin from DoneJS generator.
