+++ .editorconfig
@@ -1,0 +1,6 @@
+; Unix-style newlines
+[*]
+end_of_line = LF
+indent_style = tab
+trim_trailing_whitespace = false
+insert_final_newline = true
