+++ .jshintrc
@@ -1,0 +1,50 @@
+{
+	"globals": {
+		"steal": true,
+		"can": true,
+		"Zepto": true,
+		"QUnit": true,
+		"System": true,
+		"SimpleDOM": true,
+		"test": true,
+		"asyncTest": true,
+		"expect": true,
+		"module": true,
+		"ok": true,
+		"equal": true,
+		"notEqual": true,
+		"deepEqual": true,
+		"notDeepEqual": true,
+		"strictEqual": true,
+		"notStrictEqual": true,
+		"raises": true,
+		"start": true,
+		"stop": true,
+		"global": true
+	},
+	"curly": true,
+	"eqeqeq": true,
+	"freeze": true,
+	"indent": 2,
+	"latedef": true,
+	"noarg": true,
+	"undef": true,
+	"unused": true,
+	"trailing": true,
+	"maxdepth": 4,
+	"boss" : true,
+	"eqnull": true,
+	"evil": true,
+	"loopfunc": true,
+	"smarttabs": true,
+	"maxerr" : 200,
+	"jquery": true,
+	"dojo": true,
+	"mootools": true,
+	"yui": true,
+	"browser": true,
+	"phantom": true,
+	"rhino": true,
+	"node": true,
+	"esnext": true
+}
