+++ src/canjs-feathers.js
@@ -1,0 +1,3 @@
+export default function() {
+  return `This is the canjs-feathers plugin`;
+}
