+++ src/canjs-feathers.md
@@ -1,0 +1,5 @@
+@page canjs-feathers
+
+# canjs-feathers
+
+A set of utils for using CanJS with Feathers on the client.
