+++ test.old/test.html
@@ -1,0 +1,3 @@
+<title>canjs-feathers</title>
+<script src="../node_modules/steal/steal.js" main="test/test"></script>
+<div id="qunit-fixture"></div>
